+++ Runtime/Theme/SerieTheme.cs
@@ -10,7 +10,6 @@ namespace XCharts.Runtime
         [SerializeField] protected float m_LineSymbolSize;
         [SerializeField] protected float m_ScatterSymbolSize;
         [SerializeField] protected float m_PieTooltipExtraRadius;
-        [SerializeField] protected float m_SelectedRate = 1.3f;
         [SerializeField] protected float m_PieSelectedOffset;
         [SerializeField] protected Color32 m_CandlestickColor = new Color32(235, 84, 84, 255);
         [SerializeField] protected Color32 m_CandlestickColor0 = new Color32(71, 178, 98, 255);
@@ -37,11 +36,6 @@ namespace XCharts.Runtime
             set { if (PropertyUtil.SetStruct(ref m_LineSymbolSize, value)) SetVerticesDirty(); }
         }
         /// <summary>
-        /// the selected symbol size of line serie.
-        /// |折线图Symbol在被选中状态时的大小。
-        /// </summary>
-        public float lineSymbolSelectedSize { get { return lineSymbolSize * selectedRate; } }
-        /// <summary>
         /// the symbol size of scatter serie.
         /// |散点图的Symbol大小。
         /// </summary>
@@ -50,21 +44,6 @@ namespace XCharts.Runtime
             get { return m_ScatterSymbolSize; }
             set { if (PropertyUtil.SetStruct(ref m_ScatterSymbolSize, value)) SetVerticesDirty(); }
         }
-        /// <summary>
-        /// the selected symbol size of scatter serie.
-        /// |散点图的Symbol在被选中状态时的大小。
-        /// </summary>
-        public float scatterSymbolSelectedSize { get { return scatterSymbolSize * selectedRate; } }
-        /// <summary>
-        /// the rate of symbol size of line or scatter serie.
-        /// |折线图或散点图在被选中时的放大倍数。
-        /// </summary>
-        public float selectedRate
-        {
-            get { return m_SelectedRate; }
-            set { if (PropertyUtil.SetStruct(ref m_SelectedRate, value)) SetVerticesDirty(); }
-        }
-
         /// <summary>
         /// the extra radius of pie when actived by tooltip.
         /// |饼图鼠标移到高亮时的额外半径
@@ -130,7 +109,6 @@ namespace XCharts.Runtime
             m_LineWidth = theme.lineWidth;
             m_LineSymbolSize = theme.lineSymbolSize;
             m_ScatterSymbolSize = theme.scatterSymbolSize;
-            selectedRate = theme.selectedRate;
             m_PieTooltipExtraRadius = theme.pieTooltipExtraRadius;
             m_PieSelectedOffset = theme.pieSelectedOffset;
             m_CandlestickColor = theme.candlestickColor;